# rumicoach-backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: rumi.coach API
  version: 1.0.0
servers:
  - url: /v1
paths:
...
  /me/transactions:
    get:
      summary: Get the user's balance transaction history (newest first)
      description: >
        The minutes bank's ledger, bank-statement style: signed credits and debits with
        the balance after each row. With grouped=true, consecutive message_usage debits
        falling on the same calendar day (X-Timezone header decides the day boundary,
        default UTC) collapse into one row carrying their summed amount, the newest
        row's id/balanceAfter/createdAt, plus day and messageCount. A run is broken by
        any interleaved non-message row, so the running balance stays arithmetically
        consistent between adjacent rows.
      tags:
        - Users
      operationId: GetTransactions
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
        - name: grouped
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Collapse same-day runs of message_usage debits into one row per run.
      responses:
        '200':
          description: Paginated list of balance transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceTransactionPaginatedResponse'
        '401':
          description: Unauthorized
  /me/usage:
    get:
      summary: Get the user's usage history (billable sessions and companion messages, newest first)
      deprecated: true
      description: >
        DEPRECATED: the app's Subscription & usage screen now renders the statement view
        (GET /me/transactions?grouped=true) instead. Kept only for app versions already
        in the field; delete the endpoint and internal/services/usage once the
        statement-based release is the minimum supported app version.

        Merged usage feed for the Subscription & usage screen. Session entries come from
        the balance ledger's session rows — session_usage debits plus the zero-amount
        session_free rows that mark the free opening pair, flagged `free: true` and shown
        at their real duration. Message entries are companion-channel messages sent by
        Rumi, grouped per calendar day (X-Timezone header decides the day boundary,
        default UTC), each charged from the ledger's message_usage debits. This is usage,
        not session history — no transcripts or summaries.
      tags:
        - Users
      operationId: GetUsage
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated usage history plus lifetime totals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageHistoryResponse'
        '401':
          description: Unauthorized
...
  /admin/leads:
    get:
      summary: List business leads
      tags:
        - Admin
      security:
        - BearerAuth: []
      operationId: AdminListLeads
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: state
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: language
          schema:
            type: string
        - in: query
          name: size
          schema:
            type: string
        - in: query
          name: startDate
          schema:
            type: string
            format: date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadListResponse'
  /admin/leads/{id}:
    patch:
      summary: Update a lead
      tags:
        - Admin
      security:
        - BearerAuth: []
      operationId: AdminUpdateLead
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLeadRequest'
      responses:
        '200':
          description: Lead updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadAdmin'
...
  /admin/users/{id}/credits:
    post:
      summary: Credit session minutes to a user (manual, until a payment provider exists)
      tags:
        - Admin
      operationId: PostAdminUsersIdCredits
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreditRequest'
      responses:
        '201':
          description: Credit applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceTransaction'
        '400':
          description: Invalid amount or type
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (Admin only)
        '404':
          description: User not found
...
  /chat/sessions/{sessionID}/feedback:
    post:
      summary: Submit user feedback for a completed coaching session
      tags:
        - Chat
      operationId: SubmitSessionFeedback
      security:
        - BearerAuth: []
      parameters:
        - name: sessionID
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitFeedbackRequest'
      responses:
        '200':
          description: Feedback submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                required:
                  - status
                  - message
        '400':
          description: Bad request / validation error
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (session belongs to another user)
        '404':
          description: Session not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LeadAdmin:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        company:
          type: string
        country:
          type: string
        size:
          type: string
        message:
          type: string
        origin:
          type: string
        language:
          type: string
        state:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    LeadListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LeadAdmin'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    UpdateLeadRequest:
      type: object
      properties:
        state:
          type: string
      required:
        - state
...
    BalanceTransaction:
      type: object
      properties:
        id:
          type: string
          description: For a grouped message row, the id of the newest debit in the run.
        type:
          type: string
          enum: [subscription, top_up, session_usage, session_free, message_usage, refund]
        amountSeconds:
          type: integer
          format: int64
          description: >-
            Signed: positive = credit, negative = debit. Always 0 for session_free —
            the row exists so the history accounts for the session, not to move money.
            For a grouped message row, the run's debits summed.
        balanceAfter:
          type: integer
          format: int64
          description: The balance after this row applied (after the newest debit, for a grouped message row).
        sessionId:
          type: string
          description: Communication session charged by a session_usage debit (or recorded by a session_free row)
        sessionType:
          type: string
          description: Type of the session (if applicable)
        product:
          type: string
          description: Plan/package label for purchase credits and refunds
        description:
          type: string
        createdAt:
          type: string
          format: date-time
          description: When this row applied (the newest debit's time, for a grouped message row).
        day:
          type: string
          description: "Grouped message rows only: the calendar day of the run, YYYY-MM-DD in the request's X-Timezone."
        messageCount:
          type: integer
          description: "Grouped message rows only: how many charged replies the run collapses."
        plan:
          type: string
          enum: [monthly, annual]
          description: >-
            Billing cadence of a subscription credit, derived server-side from the
            product id so clients translate an enum instead of parsing store slugs.
            Subscription rows only, and absent when the product names neither cadence.
    BalanceTransactionPaginatedResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BalanceTransaction'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    UsageEntry:
      type: object
      required:
        - kind
        - occurredAt
        - seconds
      properties:
        kind:
          type: string
          enum: [session, messages]
        occurredAt:
          type: string
          format: date-time
          description: "Session end for sessions; the day's latest message for message groups. Sort key of the feed."
        seconds:
          type: integer
          format: int64
          description: >-
            How long this entry ran, in seconds (always positive). For a free session
            this is the real duration even though nothing was charged — see `free`, and
            note it is therefore NOT always what came off the balance.
        free:
          type: boolean
          description: >-
            True when this session was one of the free introductory ones and cost the
            user nothing. Session entries only. Free sessions are shown rather than
            hidden so the history accounts for every session the user actually had;
            they contribute to sessionCount and freeSessionSeconds, never to
            sessionSeconds.
        sessionId:
          type: string
          description: Communication session this entry billed (session entries only)
        sessionType:
          type: string
          description: Type of the billed session (session entries only)
        date:
          type: string
          description: "Day the messages were sent (YYYY-MM-DD in the request timezone) — message entries only"
        messageCount:
          type: integer
          description: >-
            Charged replies in this day group (message entries only). Proactive nudges,
            templates and system notices cost the user nothing and are not listed.
    UsageTotals:
      type: object
      properties:
        totalSeconds:
          type: integer
          format: int64
          description: sessionSeconds + messageSeconds
        sessionSeconds:
          type: integer
          format: int64
          description: Seconds actually charged to the balance; excludes free sessions.
        freeSessionSeconds:
          type: integer
          format: int64
          description: Seconds spent in free introductory sessions, charged to nobody.
        sessionCount:
          type: integer
          description: Every session in the feed, free ones included.
        messageSeconds:
          type: integer
          format: int64
        messageCount:
          type: integer
          description: Total charged replies (not day groups); excludes anything sent unprompted.
    UsageHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UsageEntry'
        totals:
          $ref: '#/components/schemas/UsageTotals'
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
    AdminCreditRequest:
      type: object
      required:
        - type
      properties:
        amountSeconds:
          type: integer
          format: int64
          description: Exactly one of amountSeconds/amountMinutes must be set and positive
        amountMinutes:
          type: integer
          description: Convenience alternative to amountSeconds (credited as minutes × 60)
        type:
          type: string
          enum: [subscription, top_up]
        product:
          type: string
          description: Plan/package label (e.g. "monthly_120", "topup_60")
        description:
          type: string
...
    PaginationInfo:
      type: object
      properties:
        currentPage:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        itemsPerPage:
          type: integer
...
    SubmitFeedbackRequest:
      type: object
      properties:
        evaluation:
          type: number
          format: double
          description: "Numeric evaluation rating (decimal, 0 to 10)"
        feedback:
          type: string
          description: "Written feedback text notes"
      required:
        - evaluation
...